+++ src/main/java/com/goormthon/bookduchilseong/domain/userbookclub/repository/UserBookClubRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.goormthon.bookduchilseong.domain.bookclub.entity.BookClub;
-import com.goormthon.bookduchilseong.domain.bookclub.entity.User;
+import com.goormthon.bookduchilseong.domain.user.entity.User;
 import com.goormthon.bookduchilseong.domain.userbookclub.entity.UserBookClub;
 
 public interface UserBookClubRepository extends JpaRepository<UserBookClub, Long> {
